+++ app/src/main/java/com/example/recipeapp/presentation/MainActivity.kt
@@ -1,0 +1,28 @@
+package com.example.recipeapp.presentation
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.ui.Modifier
+import com.example.recipeapp.presentation.account.composable.AccountContent
+import com.example.recipeapp.ui.theme.RecipeAppTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            RecipeAppTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    AccountContent()
+                }
+            }
+        }
+    }
+}
