+++ app/src/main/java/com/example/recipeapp/presentation/common/composable/SearchBarItem.kt
@@ -1,0 +1,60 @@
+package com.example.recipeapp.presentation.common.composable
+
+import android.content.res.Configuration
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.FilterList
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.SearchBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.example.recipeapp.ui.theme.RecipeAppTheme
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun SearchBarItem(
+) {
+    SearchBar(
+        query = "searching",
+        onQueryChange = {},
+        onSearch = {},
+        active = false,
+        onActiveChange = {},
+        trailingIcon = {
+            IconButton(
+                onClick = {}
+            ) {
+                Icon(
+                    imageVector = Icons.Filled.FilterList,
+                    contentDescription = "Filter"
+                )
+            }
+        },
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(bottom = 8.dp)
+            .padding(horizontal = 16.dp)
+    ) {
+
+    }
+}
+
+@Preview(
+    name = "Light Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_NO
+)
+@Preview(
+    name = "Dark Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+fun SearchBarItemPreview() {
+    RecipeAppTheme {
+        SearchBarItem()
+    }
+}
