+++ app/src/main/java/com/example/recipeapp/presentation/home/composable/HomeCategoryItem.kt
@@ -1,0 +1,76 @@
+package com.example.recipeapp.presentation.home.composable
+
+import android.content.res.Configuration
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Image
+import androidx.compose.material3.Card
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.example.recipeapp.ui.theme.RecipeAppTheme
+
+@Composable
+fun HomeCategoryItem(
+) {
+    Card(
+        shape = RoundedCornerShape(10.dp),
+        border = BorderStroke(1.dp, MaterialTheme.colorScheme.secondary)
+    ) {
+        Column(
+            modifier = Modifier
+                .padding(8.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            Row() {
+                Icon(
+                    imageVector = Icons.Filled.Image,
+                    contentDescription = null,
+                )
+
+                Icon(
+                    imageVector = Icons.Filled.Image,
+                    contentDescription = null
+                )
+            }
+
+            Row() {
+                Icon(
+                    imageVector = Icons.Filled.Image,
+                    contentDescription = null
+                )
+
+                Icon(
+                    imageVector = Icons.Filled.Image,
+                    contentDescription = null
+                )
+            }
+
+            Text(text = "Category")
+        }
+    }
+}
+
+@Preview(
+    name = "Light Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_NO
+)
+@Preview(
+    name = "Dark Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+fun HomeCategoryItemPreview() {
+    RecipeAppTheme {
+        HomeCategoryItem()
+    }
+}
