+++ app/src/main/java/com/example/recipeapp/presentation/home/composable/HomeContent.kt
@@ -1,0 +1,84 @@
+package com.example.recipeapp.presentation.home.composable
+
+import android.content.res.Configuration
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.recipeapp.presentation.common.composable.RecipeItem
+import com.example.recipeapp.presentation.common.composable.SearchBarItem
+import com.example.recipeapp.ui.theme.RecipeAppTheme
+
+@Composable
+fun HomeContent() {
+    Scaffold(
+        modifier = Modifier
+            .fillMaxSize()
+    ) { paddingValues ->
+        Column(
+            modifier = Modifier
+                .padding(paddingValues)
+        ) {
+            SearchBarItem()
+
+            TopCategoriesSection()
+
+            Card(
+                shape = RoundedCornerShape(10.dp),
+                modifier = Modifier
+                    .fillMaxSize()
+            ) {
+                Column(
+                    modifier = Modifier
+                        .padding(8.dp)
+                        .padding(top = 16.dp)
+                ) {
+                    Text(
+                        text = "Recipes",
+                        fontWeight = FontWeight.SemiBold,
+                        fontSize = 24.sp,
+                        modifier = Modifier
+                            .padding(bottom = 8.dp)
+                    )
+
+                    LazyVerticalGrid(
+                        columns = GridCells.Fixed(2),
+                        horizontalArrangement = Arrangement.spacedBy(16.dp),
+                        verticalArrangement = Arrangement.spacedBy(16.dp)
+                    ) {
+                        items(10) {
+                            RecipeItem()
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Preview(
+    name = "Light Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_NO
+)
+@Preview(
+    name = "Dark Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+fun HomeContentPreview() {
+    RecipeAppTheme {
+        HomeContent()
+    }
+}
