+++ app/src/main/java/com/example/recipeapp/presentation/home/composable/TopCategoriesSection.kt
@@ -1,0 +1,68 @@
+package com.example.recipeapp.presentation.home.composable
+
+import android.content.res.Configuration
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.recipeapp.ui.theme.RecipeAppTheme
+
+@Composable
+fun TopCategoriesSection(
+) {
+    Card(
+        shape = RoundedCornerShape(10.dp),
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(bottom = 8.dp)
+    ) {
+        Column(
+            modifier = Modifier
+                .padding(8.dp)
+                .padding(top = 16.dp)
+        ) {
+            Text(
+                text = "Top Categories",
+                fontWeight = FontWeight.SemiBold,
+                fontSize = 24.sp
+            )
+
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(top = 8.dp),
+                horizontalArrangement = Arrangement.SpaceBetween
+            ) {
+                HomeCategoryItem()
+                HomeCategoryItem()
+                HomeCategoryItem()
+                HomeCategoryItem()
+            }
+        }
+    }
+}
+
+@Preview(
+    name = "Light Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_NO
+)
+@Preview(
+    name = "Dark Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+fun HomeTopCategoriesSectionPreview() {
+    RecipeAppTheme {
+        TopCategoriesSection()
+    }
+}
