+++ app/src/main/java/com/example/recipeapp/presentation/recipe_details/composable/RecipeDetailsContent.kt
@@ -1,0 +1,219 @@
+package com.example.recipeapp.presentation.recipe_details.composable
+
+import android.content.res.Configuration
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material.icons.filled.HorizontalRule
+import androidx.compose.material.icons.outlined.Bookmark
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.SecondaryTabRow
+import androidx.compose.material3.Tab
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import coil.compose.AsyncImage
+import coil.request.ImageRequest
+import com.example.recipeapp.R
+import com.example.recipeapp.presentation.common.composable.RecipeIngredientItem
+import com.example.recipeapp.ui.theme.RecipeAppTheme
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun RecipeDetailsContent() {
+
+    var state by remember { mutableStateOf(0) }
+    val titles = listOf("Ingredients", "Description")
+
+    Scaffold(
+        modifier = Modifier
+            .fillMaxSize(),
+        topBar = {
+            TopAppBar(
+                title = {},
+                navigationIcon = {
+                    IconButton(
+                        onClick = {}
+                    ) {
+                        Icon(
+                            imageVector = Icons.AutoMirrored.Filled.ArrowBack,
+                            contentDescription = "Back button"
+                        )
+                    }
+                },
+                actions = {
+                    IconButton(
+                        onClick = {}
+                    ) {
+                        Icon(
+                            imageVector = Icons.Outlined.Bookmark,
+                            contentDescription = "Bookmark recipe button"
+                        )
+                    }
+                }
+            )
+        }
+    ) { paddingValues ->
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(paddingValues)
+        ) {
+            Text(
+                text = "Recipe name that is very long",
+                fontWeight = FontWeight.SemiBold,
+                fontSize = 24.sp,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(bottom = 8.dp)
+                    .padding(horizontal = 16.dp)
+            )
+
+            Box(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(180.dp),
+                contentAlignment = Alignment.Center
+            ) {
+                AsyncImage(
+                    model = ImageRequest
+                        .Builder(LocalContext.current)
+                        .data(R.drawable.ic_image)
+                        .crossfade(true)
+                        .placeholder(R.drawable.ic_image)
+                        .build(),
+                    contentDescription = "IMAGE",
+                    fallback = painterResource(R.drawable.ic_image),
+                    error = painterResource(R.drawable.ic_image),
+                    contentScale = ContentScale.FillWidth,
+                    modifier = Modifier
+                        .clickable {}
+                )
+            }
+
+            Column() {
+                SecondaryTabRow(selectedTabIndex = state) {
+                    titles.forEachIndexed { index, title ->
+                        Tab(
+                            selected = state == index,
+                            onClick = { state = index },
+                            text = {
+                                Text(
+                                    text = title,
+                                    maxLines = 2,
+                                    overflow = TextOverflow.Ellipsis
+                                )
+                            }
+                        )
+                    }
+                }
+            }
+
+            if (state == 0) {
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(horizontal = 16.dp),
+                    verticalAlignment = Alignment.CenterVertically
+                ) {
+                    IconButton(
+                        onClick = {},
+                    ) {
+                        Icon(
+                            imageVector = Icons.Default.HorizontalRule,
+                            contentDescription = "Less button",
+                            modifier = Modifier
+                                .border(1.dp, MaterialTheme.colorScheme.secondary, RoundedCornerShape(4.dp))
+                        )
+                    }
+
+                    Text(
+                        text = "1234 Servings",
+                        fontWeight = FontWeight.Light,
+                        fontSize = 12.sp,
+                        modifier = Modifier
+                            .padding(start = 16.dp)
+                            .padding(end = 16.dp)
+                    )
+
+                    IconButton(
+                        onClick = {}
+                    ) {
+                        Icon(
+                            imageVector = Icons.Default.Add,
+                            contentDescription = "More button",
+                            modifier = Modifier
+                                .border(1.dp, MaterialTheme.colorScheme.secondary, RoundedCornerShape(4.dp))
+                        )
+                    }
+                }
+
+                Column(
+                    modifier = Modifier
+                        .padding(bottom = 20.dp)
+                        .padding(horizontal = 16.dp)
+                ) {
+                    for (i in 1..7) {
+                        RecipeIngredientItem()
+
+                        if (i != 7) {
+                            HorizontalDivider()
+                        }
+                    }
+                }
+            }
+            else {
+                Text(
+                    text = "This is a description of a recipe.\nNew line\nAnother line",
+                    modifier = Modifier
+                        .fillMaxSize()
+                        .padding(16.dp, 24.dp)
+                )
+            }
+        }
+    }
+}
+
+@Preview(
+    name = "Light Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_NO
+)
+@Preview(
+    name = "Dark Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+fun RecipeDetailsContentPreview() {
+    RecipeAppTheme {
+        RecipeDetailsContent()
+    }
+}
