+++ app/src/main/java/com/example/recipeapp/presentation/shopping_list/composable/ShoppingListCategoryItem.kt
@@ -1,0 +1,64 @@
+package com.example.recipeapp.presentation.shopping_list.composable
+
+import android.content.res.Configuration
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Card
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.example.recipeapp.ui.theme.RecipeAppTheme
+
+@Composable
+fun ShoppingListCategoryItem() {
+    Card(
+        shape = RoundedCornerShape(10.dp),
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(bottom = 8.dp)
+    ) {
+        Column(
+            modifier = Modifier
+                .padding(8.dp)
+        ) {
+            Text(
+                text = "Category Name",
+                fontWeight = FontWeight.SemiBold,
+                modifier = Modifier
+                    .padding(bottom = 8.dp)
+                    .padding(start = 16.dp)
+            )
+
+            Column() {
+                for (i in 1..3) {
+                    ShoppingListIngredientItem()
+
+                    if (i != 3) {
+                        HorizontalDivider()
+                    }
+                }
+            }
+        }
+    }
+}
+
+@Preview(
+    name = "Light Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_NO
+)
+@Preview(
+    name = "Dark Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+fun ShoppingListCategoryItemPreview() {
+    RecipeAppTheme {
+        ShoppingListCategoryItem()
+    }
+}
