+++ app/src/main/java/com/example/recipeapp/presentation/shopping_list/composable/ShoppingListContent.kt
@@ -1,0 +1,82 @@
+package com.example.recipeapp.presentation.shopping_list.composable
+
+import android.content.res.Configuration
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.FloatingActionButton
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.recipeapp.ui.theme.RecipeAppTheme
+
+@Composable
+fun ShoppingListContent() {
+    Scaffold(
+        modifier = Modifier
+            .fillMaxSize(),
+        floatingActionButton = {
+            FloatingActionButton(
+                onClick = {}
+            ) {
+                Icon(
+                    Icons.Default.Add,
+                    contentDescription = "Add"
+                )
+            }
+        }
+    ) { paddingValues ->
+        Column(
+            modifier = Modifier
+                .padding(paddingValues)
+        ) {
+            Text(
+                text = "Shopping List",
+                fontWeight = FontWeight.SemiBold,
+                fontSize = 24.sp,
+                modifier = Modifier
+                    .padding(bottom = 8.dp)
+                    .padding(top = 8.dp)
+                    .padding(start = 16.dp)
+            )
+
+            Text(
+                text = "123 items",
+                fontWeight = FontWeight.Light,
+                modifier = Modifier
+                    .padding(bottom = 8.dp)
+                    .padding(start = 16.dp)
+            )
+
+            LazyColumn() {
+                items(5) {
+                    ShoppingListCategoryItem()
+                }
+            }
+        }
+    }
+}
+
+@Preview(
+    name = "Light Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_NO
+)
+@Preview(
+    name = "Dark Mode",
+    uiMode = Configuration.UI_MODE_NIGHT_YES
+)
+@Composable
+fun ShoppingListContentPreview() {
+    RecipeAppTheme {
+        ShoppingListContent()
+    }
+}
